requirements.txt
matplotlib==3.5.2
numpy==1.22.4
omegaconf==2.1.2
torch==1.11.0
Pillow==9.1.1
hydra-core==1.0.3